ReservationsContainer 를 구현하라
예약 현황 상태를 관리하는 컨테이너 컴포넌트를 구현하였습니다.
